+++ .github/CODEOWNERS
@@ -174,6 +174,7 @@
 # BEGIN COMPONENTS
 /comp @DataDog/agent-shared-components
 /comp/core @DataDog/agent-shared-components
+/comp/flare @DataDog/agent-shared-components
 # END COMPONENTS
 
 # pkg
