+++ comp/flare/flare/component.go
