+++ comp/flare/flare/helpers/builder.go
