+++ comp/flare/flare/helpers/perm_info.go
