+++ comp/flare/flare/helpers/perm_info_nix.go
