+++ comp/flare/flare/helpers/perm_info_win.go
