+++ comp/flare/flare/helpers/perm_info_win_test.go
