+++ pkg/flare/archive_linux.go
@@ -15,7 +15,7 @@ import (
 	"regexp"
 	"syscall"
 
-	flarehelpers "github.com/DataDog/datadog-agent/comp/core/flare/helpers"
+	flarehelpers "github.com/DataDog/datadog-agent/comp/flare/flare/helpers"
 )
 
 func getLinuxKernelSymbols(fb flarehelpers.FlareBuilder) error {
